# docs/swagger.yaml
schemes:
  - http
  - https
swagger: '2.0'
info:
  description: REST API for the LMS platform covering auth, courses, enrollments, assessments, payments, and more.
  title: LMS Backend API
  contact: {}
  version: 1.0.0
host: localhost:8080
basePath: /
paths:
...
  /v1/admin/points/config:
    patch:
      security:
        - BearerAuth: []
      description: Admin updates the platform-wide points configuration for video completions, quiz passes, and perfect score bonuses
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - points
      summary: Update points configuration
      parameters:
        - description: Points configuration update
          name: body
          in: body
          schema:
            type: object
            properties:
              bonus_points_perfect_score:
                type: integer
              points_per_quiz_pass:
                type: integer
              points_per_video:
                type: integer
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/points.PointsConfigResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/admin/rbac/roles:
    get:
      security:
        - BearerAuth: []
      description: Returns all roles with their current permission sets
      produces:
        - application/json
      tags:
        - rbac
      summary: List roles
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/rbac.ListRolesResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/admin/rbac/roles/{roleId}:
    patch:
      security:
        - BearerAuth: []
      description: Replaces the permission set for the specified role
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - rbac
      summary: Update role permissions
      parameters:
        - type: string
          description: Role ID
          name: roleId
          in: path
          required: true
        - description: Permissions payload
          name: body
          in: body
          required: true
          schema:
            type: object
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/rbac.UpdateRolePermissionsResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
...
  /v1/leaderboard:
    get:
      security:
        - BearerAuth: []
      description: Returns ranked leaderboard entries for the specified period. Opted-out students are shown as Anonymous to other users.
      produces:
        - application/json
      tags:
        - points
      summary: Get leaderboard
      parameters:
        - enum:
            - weekly
            - alltime
          type: string
          default: alltime
          description: 'Leaderboard period: weekly or alltime'
          name: period
          in: query
        - type: integer
          default: 1
          description: Page number
          name: page
          in: query
        - maximum: 100
          type: integer
          default: 20
          description: Items per page (max 100)
          name: limit
          in: query
        - type: string
          description: Filter by course ID (UUID)
          name: course_id
          in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/points.LeaderboardResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
...
  /v1/student/leaderboard/opt-out:
    patch:
      security:
        - BearerAuth: []
      description: Allows the authenticated student to opt in or out of the public leaderboard
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - points
      summary: Toggle leaderboard opt-out
      parameters:
        - description: Opt-out toggle request
          name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              opt_out:
                type: boolean
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/points.ToggleLeaderboardOptOutResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/student/points:
    get:
      security:
        - BearerAuth: []
      description: Returns total_points, points_today, points_this_week, daily_breakdown_today, global_rank, and weekly_rank for the authenticated student
      produces:
        - application/json
      tags:
        - points
      summary: Get student points summary
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/points.StudentPointsResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/student/points/history:
    get:
      security:
        - BearerAuth: []
      description: Returns a paginated log of point-earning events for the authenticated student
      produces:
        - application/json
      tags:
        - points
      summary: Get points history
      parameters:
        - type: integer
          default: 1
          description: Page number
          name: page
          in: query
        - maximum: 100
          type: integer
          default: 20
          description: Items per page (max 100)
          name: limit
          in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/points.PointsHistoryResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
...
definitions:
...
  handlers.ErrorResponse:
    type: object
    properties:
      error:
        type: object
        properties:
          code:
            type: string
          message:
            type: string
  handlers.ValidationDetail:
    type: object
    properties:
      field:
        type: string
      message:
        type: string
  handlers.ValidationErrorResponse:
    type: object
    properties:
      error:
        type: object
        properties:
          code:
            type: string
          details:
            type: array
            items:
              $ref: '#/definitions/handlers.ValidationDetail'
          message:
            type: string
...
  points.DailyBreakdownEntry:
    type: object
    properties:
      bonus_points:
        type: integer
      earned_at:
        type: string
      points:
        type: integer
      source_title:
        type: string
      type:
        type: string
  points.LeaderboardEntry:
    type: object
    properties:
      display_name:
        type: string
      rank:
        type: integer
      score:
        type: number
      student_id:
        type: string
  points.LeaderboardResponse:
    type: object
    properties:
      entries:
        type: array
        items:
          $ref: '#/definitions/points.LeaderboardEntry'
      period:
        type: string
  points.PaginationMeta:
    type: object
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
  points.PointEventResponse:
    type: object
    properties:
      bonus_points:
        type: integer
      earned_at:
        type: string
      id:
        type: string
      points:
        type: integer
      source_id:
        type: string
      source_title:
        type: string
      type:
        type: string
  points.PointsConfigResponse:
    type: object
    properties:
      bonus_points_perfect_score:
        type: integer
      points_per_quiz_pass:
        type: integer
      points_per_video:
        type: integer
  points.PointsHistoryResponse:
    type: object
    properties:
      events:
        type: array
        items:
          $ref: '#/definitions/points.PointEventResponse'
      meta:
        $ref: '#/definitions/points.PaginationMeta'
  points.StudentPointsResponse:
    type: object
    properties:
      daily_breakdown_today:
        type: array
        items:
          $ref: '#/definitions/points.DailyBreakdownEntry'
      global_rank:
        type: integer
      points_this_week:
        type: integer
      points_today:
        type: integer
      total_points:
        type: integer
      weekly_rank:
        type: integer
  points.ToggleLeaderboardOptOutResponse:
    type: object
    properties:
      opted_out:
        type: boolean
  rbac.ListRolesResponse:
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/rbac.RoleResponse'
  rbac.PermissionResponse:
    type: object
    properties:
      action:
        type: string
      resource:
        type: string
  rbac.RoleResponse:
    type: object
    properties:
      permissions:
        type: array
        items:
          $ref: '#/definitions/rbac.PermissionResponse'
      role:
        type: string
  rbac.UpdateRolePermissionsResponse:
    type: object
    properties:
      permissions:
        type: array
        items:
          $ref: '#/definitions/rbac.PermissionResponse'
      role:
        type: string
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header